+++ BeSafeWebApp.Contracts/Interfaces/BusinessLogicLayers/IMasterItemBusinessLogic.cs
@@ -7,8 +7,8 @@ namespace BeSafeWebApp.Contracts.Interfaces
     public interface IMasterItemBusinessLogic
     {
         Task<IList<Entities.MasterItemsSet>> GetAllMasterItems();
-        Task<IList<Entities.MasterItemsSet>> GetMasterItemsByCategoryId(int CategoryId);
-        Task<Entities.MasterItemsSet> GetMasterItemById(int id);
+        Task<IList<Entities.MasterItemsSet>> GetMasterItemsByCategoryId(long CategoryId);
+        Task<Entities.MasterItemsSet> GetMasterItemById(long id);
         Task AddMasterItem(Entities.MasterItemsSet category);
         Task UpdateMasterItem(Entities.MasterItemsSet category);
         Task DeleteMasterItem(Entities.MasterItemsSet category);
