+++ BeSafeWebApp.Contracts/Interfaces/Repositories/IMasterItemRepository.cs
@@ -6,7 +6,7 @@ namespace BeSafeWebApp.Contracts.Interfaces
 {
     public interface IMasterItemRepository : IGenericRepository<Entities.MasterItemsSet>
     {
-        Task<IList<Entities.MasterItemsSet>> GetMasterItemsByCategoryId(int CategoryId);
+        Task<IList<Entities.MasterItemsSet>> GetMasterItemsByCategoryId(long CategoryId);
 
     }
 }
